Fix redundant C-cast

+++ ra/2keyfram.cpp
@@ -321,7 +321,7 @@ void* Build_Frame(void const* dataptr, unsigned short framenumber,
       ptr =
           (char*)Add_Long_To_Pointer(dataptr, ((unsigned long)currframe << 3) +
                                                   sizeof(KeyFrameHeaderType));
-      Mem_Copy(ptr, &offset[0], (long)(SUBFRAMEOFFS * sizeof(uint32_t)));
+      Mem_Copy(ptr, &offset[0], SUBFRAMEOFFS * sizeof(uint32_t));
     }
 
     // key frame
@@ -364,7 +364,7 @@ void* Build_Frame(void const* dataptr, unsigned short framenumber,
           Mem_Copy(
               Add_Long_To_Pointer(dataptr, ((unsigned long)currframe << 3) +
                                                sizeof(KeyFrameHeaderType)),
-              &offset[0], (long)(SUBFRAMEOFFS * sizeof(uint32_t)));
+              &offset[0], SUBFRAMEOFFS * sizeof(uint32_t));
           subframe = 0;
         }
       }
