+++ ra/template.cpp
@@ -228,7 +228,7 @@ void TemplateClass::operator delete(void* ptr) {
  *=============================================================================================*/
 TemplateClass::TemplateClass(TemplateType type, CELL pos)
     : ObjectClass(RTTI_TEMPLATE, Templates.ID(this)),
-      Class(TemplateTypes.Ptr((int)type)) {
+      Class(TemplateTypes.Ptr(type)) {
   if (pos != -1) {
     Unlimbo(Cell_Coord(pos));
   }
