+++ td/ipx95.cpp
@@ -48,11 +48,11 @@
 #include "td/ipx.h"
 
 int IPX_Open_Socket(unsigned short socket) {
-  return IPX_Open_Socket95((int)socket);
+  return IPX_Open_Socket95(socket);
 }
 
 int IPX_Close_Socket(unsigned short socket) {
-  IPX_Close_Socket95((int)socket);
+  IPX_Close_Socket95(socket);
   return 0;
 }
